The older CacheEngine wont work. V2 in place. WIP
Older cache engine had a problem with updations since there was a lot of
dependency on having to bother about uniqueness ourselves since the
method was to use individual hashmaps

This method uses a tree-ish structure to maintain the hierarchy. Seems
good enough.

This is a total Half assed implementation, please excuse. Total WIP case