Getting rid of all the Generic shit!!
Turns out, im not *that* good a java programmer yet. Bloody Generic
classes. A little more time and i wouldn't need to go to the bloody
barbed, i would've pulled out all the hair on my head with my bare
hands.

With that said, it is something that must eventually be understood!