Improving testbed. Looks like decent performance!
Trying out both DFS method to create some namespaces, structures and
instances AND a BFS method to do it. Got some rather interesting
insights. There is a point in the code that randomly generates a set of
timeseries to be used with every instance. Now if i increase that to an
ungodly number like 1000000, then it has to generate that many points.
Turns out the bottleneck i was seeing yesterday WAS that. Right now i'm
looking at 1 Billion points being added in about 200 odd ms' which seems
about right to me.

Will now move on to see exactly HOW much memory does this take.

TODO
  comments

---
:+1: @raghothams @nischalhp :+1:
Neevu ilde yen maadirthiddno naanu!